toasty_core::schema

Struct Schema

pub struct Schema {
    pub app: Schema,
    pub db: Schema,
    pub mapping: Mapping,
}

The combined schema: app-level models, database-level tables, and the mapping that connects them.

Constructed with Builder and validated on creation. Immutable at runtime.

Examples

use toasty_core::Schema;

fn inspect(schema: &Schema) {
    for (id, model) in &schema.app.models {
        let table = schema.table_for(*id);
        println!("{} -> {}", model.name().snake_case(), table.name);
    }
}

Fields
app: Schema

Application-level schema.

db: Schema

Database-level schema.

mapping: Mapping

Maps the app-level schema to the db-level schema.

Implementations

impl Schema

pub fn builder() -> Builder

Returns a new Builder for constructing a Schema.

Examples

use toasty_core::Schema;

let builder = Schema::builder();

pub fn mapping_for(&self, id: impl Into<ModelId>) -> &Model

Returns the mapping for the given model.

Panics

Panics if id does not correspond to a model in the schema.

Examples

use toasty_core::Schema;

let mapping = schema.mapping_for(model_id);
println!("table: {:?}", mapping.table);

pub fn table_for(&self, id: impl Into<ModelId>) -> &Table

Returns the database table that stores the given model.

Panics

Panics if id does not correspond to a model in the schema.

Examples

let table = schema.table_for(model_id);
println!("table name: {}", table.name);

pub fn table_id_for(&self, id: impl Into<ModelId>) -> TableId

Returns the TableId for the table that stores the given model.

Panics

Panics if id does not correspond to a model in the schema.