toasty_core::stmt

Enum Statement

pub enum Statement {
    Delete(Delete),
    Insert(Insert),
    Query(Query),
    Update(Update),
}

A top-level statement in Toasty’s AST.

Each variant corresponds to one of the four fundamental database operations. A Statement is the primary input to the query engine’s compilation pipeline and the output of code generated by #[derive(Model)].

Examples

use toasty_core::stmt::{Statement, Query, Values};

let query = Query::unit();
let stmt = Statement::from(query);
assert!(stmt.is_query());
assert!(!stmt.is_insert());

Variants

Delete(Delete)

Delete one or more existing records.

Insert(Insert)

Create one or more new records.

Query(Query)

Query (read) records from the database.

Update(Update)

Update one or more existing records.

Implementations

impl Statement

pub fn assignments(&self) -> Option<&Assignments>

Returns this statement’s assignments if it is an Update.

impl Statement

pub fn condition(&self) -> Option<&Condition>

Returns a reference to this statement’s condition, if it has one and it is set. Only Update and Delete statements support conditions.

pub fn condition_mut(&mut self) -> Option<&mut Condition>

Returns a mutable reference to the statement’s condition.

Returns None for statements that do not support conditions.

pub fn condition_mut_unwrap(&mut self) -> &mut Condition

Returns a mutable reference to the statement’s condition.

Panics

Panics if the statement does not support conditions.

impl Statement

pub fn is_delete(&self) -> bool

Returns true if this statement is a Delete.

pub fn as_delete(&self) -> Option<&Delete>

Attempts to return a reference to an inner Delete.

If self is a Statement::Delete, a reference to the inner Delete is returned wrapped in Some.

Else, None is returned.

pub fn into_delete(self) -> Option<Delete>

Consumes self and attempts to return the inner Delete.

If self is a Statement::Delete, inner Delete is returned wrapped in Some.

Else, None is returned.

pub fn into_delete_unwrap(self) -> Delete

Consumes self and returns the inner Delete.

Panics

If self is not a Statement::Delete.

impl Statement

pub fn eval(&self, input: impl Input) -> Result<Value>

Evaluates this statement using the provided Input for argument resolution. Only Query statements are supported.

Errors

Returns an error for non-Query statements, or if evaluation of any sub-expression fails.

pub fn eval_const(&self) -> Result<Value>

Evaluates this statement as a constant expression (no external input).

impl Statement

pub fn filter(&self) -> Option<&Filter>

Returns a reference to this statement’s filter, if it has one.

Returns None for INSERT statements.

pub fn filter_unwrap(&self) -> &Filter

Returns a reference to this statement’s filter.

Panics

Panics if the statement has no filter.

pub fn filter_or_default(&self) -> &Filter

Returns this statement’s filter, or Filter::ALL if it has none.

pub fn filter_mut(&mut self) -> Option<&mut Filter>

Returns a mutable reference to the statement’s filter.

Returns None for statements that do not support filtering, such as INSERT.

pub fn filter_mut_unwrap(&mut self) -> &mut Filter

Returns a mutable reference to the statement’s filter.

Panics

Panics if the statement does not support filtering.

pub fn filter_expr_unwrap(&self) -> &Expr

Returns a reference to the filter’s inner expression.

Panics

Panics if the statement has no filter or the filter has no expression.

pub fn filter_expr_mut(&mut self) -> Option<&mut Expr>

Returns a mutable reference to the filter’s inner expression, if any.

impl Statement

pub fn is_insert(&self) -> bool

Returns true if this statement is an Insert.

pub fn as_insert(&self) -> Option<&Insert>

Attempts to return a reference to an inner Insert.

If self is a Statement::Insert, a reference to the inner Insert is returned wrapped in Some.

Else, None is returned.

pub fn into_insert(self) -> Option<Insert>

Consumes self and attempts to return the inner Insert.

If self is a Statement::Insert, inner Insert is returned wrapped in Some.

Else, None is returned.

pub fn into_insert_unwrap(self) -> Insert

Consumes self and returns the inner Insert.

Panics

If self is not a Statement::Insert.

impl Statement

pub fn is_query(&self) -> bool

Returns true if this statement is a Query.

pub fn as_query(&self) -> Option<&Query>

Attempts to return a reference to an inner Query.

If self is a Statement::Query, a reference to the inner Query is returned wrapped in Some.

Else, None is returned.

pub fn as_query_mut(&mut self) -> Option<&mut Query>

Returns a mutable reference to the inner Query, if this is a query statement.

If self is a Statement::Query, a mutable reference to the inner Query is returned wrapped in Some.

Else, None is returned.

pub fn into_query(self) -> Option<Query>

Consumes self and attempts to return the inner Query.

Returns None if self is not a Statement::Query.

pub fn into_query_unwrap(self) -> Query

Consumes self and returns the inner Query.

Panics

If self is not a Statement::Query.

impl Statement

pub fn returning(&self) -> Option<&Returning>

Returns a reference to this statement’s RETURNING clause, if present.

Returns None if the statement does not have a RETURNING clause or is a statement type that does not support RETURNING.

pub fn take_returning(&mut self) -> Option<Returning>

Take the Returning clause

pub fn set_returning(&mut self, returning: Returning)

Set the Returning clause

pub fn set_returning_project(&mut self, expr: impl Into<Expr>)

Set the Returning clause to Returning::Project containing the given expression.

pub fn set_returning_expr(&mut self, expr: impl Into<Expr>)

Set the Returning clause to Returning::Expr containing the given expression.

pub fn returning_unwrap(&self) -> &Returning

Returns a reference to this statement’s RETURNING clause.

Panics

Panics if the statement does not have a RETURNING clause.

pub fn returning_mut(&mut self) -> Option<&mut Returning>

Returns a mutable reference to this statement’s RETURNING clause, if present.

Returns None if the statement does not have a RETURNING clause or is a statement type that does not support RETURNING.

pub fn returning_mut_unwrap(&mut self) -> &mut Returning

Returns a mutable reference to this statement’s RETURNING clause.

Panics

Panics if the statement does not have a RETURNING clause.

impl Statement

pub fn query_select(&self) -> Option<&Select>

If this is a query with a SELECT body, returns a reference to that Select. Returns None otherwise.

pub fn query_select_unwrap(&self) -> &Select

Returns a reference to this statement’s inner Select.

Panics

Panics if this is not a query statement with a SELECT body.

impl Statement

pub fn is_update(&self) -> bool

Returns true if this statement is an Update.

impl Statement

pub fn name(&self) -> &str

Returns the statement variant name for logging.

pub fn substitute(&mut self, input: impl Input)

Substitutes argument placeholders in this statement with concrete values from input.

pub fn is_const(&self) -> bool

Returns true if this statement is a query whose body contains only constant values (no table references or subqueries) and has no CTEs.

pub fn as_update(&self) -> Option<&Update>

Attempts to return a reference to an inner Update.

If self is a Statement::Update, a reference to the inner Update is returned wrapped in Some.

Else, None is returned.

pub fn into_update(self) -> Option<Update>

Consumes self and attempts to return the inner Update.

If self is a Statement::Update, inner Update is returned wrapped in Some.

Else, None is returned.

pub fn is_single(&self) -> bool

Returns true if this statement expects at most one result row.

pub fn into_update_unwrap(self) -> Update

Consumes self and returns the inner Update.

Panics

If self is not a Statement::Update.