toasty_core::stmt

Struct PathFieldSet

pub struct PathFieldSet { /* private fields */ }

A set of field indices, backed by a bit set.

Used to track which fields are present in a SparseRecord or which fields are part of a type description. Supports set operations like union (|), intersection (&), and membership tests.

Examples

use toasty_core::stmt::PathFieldSet;

let mut set = PathFieldSet::new();
set.insert(0);
set.insert(2);
assert!(set.contains(0_usize));
assert!(!set.contains(1_usize));
assert_eq!(set.len(), 2);

Implementations

impl PathFieldSet

pub fn new() -> Self

Creates an empty field set.

pub fn from_slice<T>(fields: &[T]) -> Self
where for<'a> &'a T: Into<usize>,

Creates a field set from a slice of values convertible to usize.

pub fn contains(&self, val: impl Into<usize>) -> bool

Returns true if the set contains the given field index.

pub fn iter(&self) -> PathFieldSetIter<'_>

Returns an iterator over the field indices in ascending order.

pub fn is_empty(&self) -> bool

Returns true if the set contains no field indices.

pub fn len(&self) -> usize

Returns the number of field indices in the set.

pub fn insert(&mut self, val: usize)

Inserts a field index into the set.