toasty_core::stmt

Trait Project

pub trait Project {
    // Required method
    fn project(self, projection: &Projection) -> Option<Expr>;
}

Trait for types that can be projected through a Projection.

Returns the projected expression, or None if the projection is not applicable.

Required Methods

fn project(self, projection: &Projection) -> Option<Expr>

Applies the projection and returns the resulting expression.